backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-ingestion-pipeline"
version = "1.0.0"
description = "RAG Ingestion Pipeline: URL crawling → text extraction → content chunking → embedding generation → vector storage"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "qdrant-client>=1.3.0",
    "cohere>=4.0.0",
    "httpx>=0.23.0",
    "urllib3>=1.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "flake8>=5.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'